pyproject.toml
[project]
name = "academic-author-disambiguation"
version = "0.1.0"
description = "Multi-source pipeline for matching Fulbright scholars to publications across OpenAlex, Scopus, and Google Scholar"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.0",
    "requests>=2.31",
    "fuzzywuzzy[speedup]>=0.18",
    "python-Levenshtein>=0.25",
    "google-cloud-bigquery>=3.0",
    "pandas-gbq>=0.19",
    "elsapy>=0.5",
    "beautifulsoup4>=4.12",
    "crossrefapi>=1.5",
    "openpyxl>=3.1",
    "XlsxWriter>=3.1",
    "unidecode>=1.3",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["ruff", "pytest"]
